+++ Changelog.md
@@ -4,6 +4,16 @@
 
 ## CHANGELOG
 
+### 0.4.0 @ 10/22/2024
+
+#### :mega: New
+
+1. Upgrade to the new version `0.4.0`.
+
+#### :wrench: Fix
+
+1. Fix: Correct a mistake of the HTML usage in the localized version `zh-cn/docs/license`.
+
 ### 0.3.4 @ 10/20/2024
 
 #### :wrench: Fix